requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
fastapi==0.135.1
google-auth==2.49.1
google-genai==1.67.0
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
